Fixed warning: conversion from double to int

+++ modules/tracking/src/trackerKCF.cpp
@@ -176,7 +176,7 @@ namespace cv{
     if(params.descriptor != GRAY){printf("The choosen descriptor mode is not available! Please use GRAY descriptor, other descriptors will be avaiable soon.\n");return false;}//temporary, will be updated soon
     
     // create gaussian response
-    y=Mat::zeros(roi.height,roi.width,CV_64F);
+    y=Mat::zeros((int)roi.height,(int)roi.width,CV_64F);
     for(unsigned i=0;i<roi.height;i++){
       for(unsigned j=0;j<roi.width;j++){
 	y.at<double>(i,j)=(i-roi.height/2+1)*(i-roi.height/2+1)+(j-roi.width/2+1)*(j-roi.width/2+1);
@@ -391,7 +391,7 @@ namespace cv{
     for(int i=0;i<_patch.rows;i++){
       for(int j=0;j<_patch.cols;j++){
 	pixel=_patch.at<Vec3b>(i,j);
-	index=floor(pixel[2]/8)+32*floor(pixel[1]/8)+32*32*floor(pixel[0]/8);
+	index=(unsigned)(floor(pixel[2]/8)+32*floor(pixel[1]/8)+32*32*floor(pixel[0]/8));
 	
 	//copy the values
 	for(int _k=0;_k<10;_k++){
